refactor: move all logic to Cell component, render a single cell, and add interface
